refacotr: extract `Recipe` type to its own file
