+++ tooling/create-tauri-app/src/recipes/react.ts
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-import { Recipe } from '..'
 import { join } from 'path'
 // @ts-expect-error
 import scaffe from 'scaffe'
 import { shell } from '../shell'
+import { Recipe } from '../types/recipe'
 
 const afterCra = async (
   cwd: string,
